Project 3 - Reading the csv file using c++
In this project you will implement the DataFrame class along with the all the methods that are
represented in the class definition. You are required to read a csv file, do operations on the data
stored and display the queried contents from the dataframe.
DataFrame is a table of data with rows and columns – columns have names associated with them
also - we call these as headers. You may or may not have headers and this will be mentioned in
the input file(more details follow). You are going to store the given information in a csv file as a
vector of DFrow objects and answer queries from the input file.